Nagivate webpage page handeling response from servlet
Changed the coordinate system of all of the starter data to WGS84 so there would not be issues in the front end converted from projected to geographic coordinate system. Also added most of the needed functionality to the navigate page.  Points are added to map with custom icons that differ between due and current parks. each point has custom popup content and there is a layers control.